@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
     <!-- SonarCloud needs this -->
     <ProjectGuid>{1eba5dd5-58eb-4f50-be1d-f30fd9fd326d}</ProjectGuid>
   </PropertyGroup>
@@ -11,8 +11,7 @@
     </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="KubernetesClient" Version="1.5.18" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="KubernetesClient" Version="1.6.12" />
     <PackageReference Include="Serilog.Extensions.Hosting" Version="2.0.0" />
     <PackageReference Include="Serilog.Extensions.Logging" Version="2.0.2" />
     <PackageReference Include="Serilog.Sinks.Console" Version="3.1.1" />
@@ -27,7 +26,7 @@
     <CodeAnalysisRuleSet>..\Altinn3.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>bin\Debug\netcoreapp2.2\KubernetesWrapper.xml</DocumentationFile>
+    <DocumentationFile>bin\Debug\netcoreapp3.1\KubernetesWrapper.xml</DocumentationFile>
   </PropertyGroup>
 
 </Project>
